feat: add halfstar review #d4d-105

+++ backend_v2/src/main/java/at/htl/endpoints/RatingResource.java
@@ -75,7 +75,7 @@ public class RatingResource {
         // 5. Bewertung erstellen und speichern
         Review review = new Review();
         review.setService(service);
-        review.setRating(ratingRequest.getStars().doubleValue());
+        review.setRating(ratingRequest.getStars());
         review.setComment(ratingRequest.getComment());
         
         reviewRepository.persist(review);
